[extension] Move optional interfaces to extensioncapabilities
