+++ extension/extensioncapabilities/interfaces.go
@@ -1,0 +1,25 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+// Package extensioncapabilities provides interfaces that can be implemented by extensions
+// to provide additional capabilities.
+package extensioncapabilities // import "go.opentelemetry.io/collector/extension/extensioncapabilities"
+
+import (
+	"go.opentelemetry.io/collector/extension/internal"
+)
+
+// Dependent is an optional interface that can be implemented by extensions
+// that depend on other extensions and must be started only after their dependencies.
+// See https://github.com/open-telemetry/opentelemetry-collector/pull/8768 for examples.
+type Dependent = internal.Dependent
+
+// PipelineWatcher is an extra interface for Extension hosted by the OpenTelemetry
+// Collector that is to be implemented by extensions interested in changes to pipeline
+// states. Typically this will be used by extensions that change their behavior if data is
+// being ingested or not, e.g.: a k8s readiness probe.
+type PipelineWatcher = internal.PipelineWatcher
+
+// ConfigWatcher is an interface that should be implemented by an extension that
+// wishes to be notified of the Collector's effective configuration.
+type ConfigWatcher = internal.ConfigWatcher
